Removes obsolete small files
