--- packages/server/src/enums/agent-measurable-score-type.js
@@ -1,22 +1,0 @@
-// @todo: To split the file.
-
-const MARKET_TYPE = {
-  MARKET: 'MARKET',
-  NON_MARKET: 'NON_MARKET',
-};
-
-const START_AT = {
-  QUESTION_CREATION_TIME: 'QUESTION_CREATION_TIME',
-  AGENT_MEASUREMENT_CREATION_TIME: 'AGENT_MEASUREMENT_CREATION_TIME',
-};
-
-const FINAL_COMPARISON_MEASUREMENT = {
-  LAST_OBJECTIVE_MEASUREMENT: 'LAST_OBJECTIVE_MEASUREMENT',
-  LAST_AGGREGATE_MEASUREMENT: 'LAST_AGGREGATE_MEASUREMENT',
-};
-
-module.exports = {
-  MARKET_TYPE,
-  START_AT,
-  FINAL_COMPARISON_MEASUREMENT,
-};
