--- packages/server/src/enums/agent-type.js
@@ -1,8 +1,0 @@
-const AGENT_TYPE = {
-  USER: 'USER',
-  BOT: 'BOT',
-};
-
-module.exports = {
-  AGENT_TYPE,
-};
