--- packages/server/src/enums/channel-membership-roles.js
@@ -1,20 +1,0 @@
-/**
- * @todo: He is a small issue. If we have "viewers"
- * @todo: so then why then can add predictions?
- * @todo: The better to give them another
- * @todo: group name.
- *
- * @todo: But here is other thing. If we have
- * @todo: "requireVerification" then this is ok.
- */
-
-const CHANNEL_MEMBERSHIP_ROLES = {
-  ADMIN: 'ADMIN',
-  VIEWER: 'VIEWER',
-  VERIFIED: 'VERIFIED',
-  NONE: 'NONE',
-};
-
-module.exports = {
-  CHANNEL_MEMBERSHIP_ROLES,
-};
