--- packages/server/src/enums/channel-membership-type.js
@@ -1,9 +1,0 @@
-const CHANNEL_MEMBERSHIP_TYPE = {
-  ADDED_IN_APP_BY_ADMIN: 'ADDED_IN_APP_BY_ADMIN',
-  AGENT_JOINED_DIRECTLY: 'AGENT_JOINED_DIRECTLY',
-  ADDED_BY_EMAIL_BY_ADMIN: 'ADDED_BY_EMAIL_BY_ADMIN',
-};
-
-module.exports = {
-  CHANNEL_MEMBERSHIP_TYPE,
-};
