--- packages/server/src/enums/feed-item-body.js
@@ -1,12 +1,0 @@
-const FEED_ITEM_BODY = {
-  generic: 'generic',
-  measurable: 'measurable',
-  measurement: 'measurement',
-  joinedMember: 'joinedMember',
-  channel: 'channel',
-  notebook: 'notebook',
-};
-
-module.exports = {
-  FEED_ITEM_BODY,
-};
