--- packages/server/src/enums/measurable-state.js
@@ -1,10 +1,0 @@
-const MEASURABLE_STATE = {
-  OPEN: 'OPEN',
-  JUDGEMENT_PENDING: 'JUDGEMENT_PENDING',
-  JUDGED: 'JUDGED',
-  CLOSED_AS_UNRESOLVED: 'CLOSED_AS_UNRESOLVED',
-};
-
-module.exports = {
-  MEASURABLE_STATE,
-};
