--- packages/server/src/enums/measurable-value-type.js
@@ -1,9 +1,0 @@
-const MEASURABLE_VALUE_TYPE = {
-  FLOAT: 'FLOAT',
-  DATE: 'DATE',
-  PERCENTAGE: 'PERCENTAGE',
-};
-
-module.exports = {
-  MEASURABLE_VALUE_TYPE,
-};
