--- packages/server/src/enums/measurement-comment-type.js
@@ -1,9 +1,0 @@
-const MEASUREMENT_COMMENT_TYPE = {
-  GENERIC: 'GENERIC',
-  QUESTION_FEEDBACK: 'QUESTION_FEEDBACK',
-  UPDATE: 'UPDATE',
-};
-
-module.exports = {
-  MEASUREMENT_COMMENT_TYPE,
-};
