--- packages/server/src/enums/measurement-competitor-type.js
@@ -1,11 +1,0 @@
-const MEASUREMENT_COMPETITOR_TYPE = {
-  OBJECTIVE: 'OBJECTIVE',
-  COMPETITIVE: 'COMPETITIVE',
-  AGGREGATION: 'AGGREGATION',
-  UNRESOLVED: 'UNRESOLVED',
-  COMMENT: 'COMMENT',
-};
-
-module.exports = {
-  MEASUREMENT_COMPETITOR_TYPE,
-};
