--- packages/server/src/enums/measurement-unresolvable-resolutions.js
@@ -1,10 +1,0 @@
-const UNRESOLVABLE_RESOLUTIONS = {
-  AMBIGUOUS: 'AMBIGUOUS',
-  RESULT_NOT_AVAILABLE: 'RESULT_NOT_AVAILABLE',
-  FALSE_CONDITIONAL: 'FALSE_CONDITIONAL',
-  OTHER: 'OTHER',
-};
-
-module.exports = {
-  UNRESOLVABLE_RESOLUTIONS,
-};
