--- packages/server/src/enums/measurement-value.js
@@ -1,13 +1,0 @@
-const MEASUREMENT_VALUE = {
-  floatCdf: 'floatCdf',
-  floatPoint: 'floatPoint',
-  percentage: 'percentage',
-  binary: 'binary',
-  unresolvableResolution: 'unresolvableResolution',
-  comment: 'comment',
-  none: 'none',
-};
-
-module.exports = {
-  MEASUREMENT_VALUE,
-};
