--- packages/server/src/enums/notification-error-reason.js
@@ -1,10 +1,0 @@
-const NOTIFICATION_ERROR_REASON = {
-  INTERNAL_ERROR: 'INTERNAL_ERROR',
-  EXTERNAL_ERROR: 'EXTERNAL_ERROR',
-  PREFERENCES_ERROR: 'PREFERENCES_ERROR',
-  EMAIL_ADDRESS_ERROR: 'EMAIL_ADDRESS_ERROR',
-};
-
-module.exports = {
-  NOTIFICATION_ERROR_REASON,
-};
