--- packages/server/src/enums/notification-type.js
@@ -1,9 +1,0 @@
-const NOTIFICATION_TYPE = {
-  EMAIL: 'EMAIL',
-  PUSH: 'PUSH',
-  WEB_PUSH: 'WEB_PUSH',
-};
-
-module.exports = {
-  NOTIFICATION_TYPE,
-};
