--- packages/server/src/enums/template-name.js
@@ -1,23 +1,0 @@
-const TEMPLATE_NAME = {
-  MEASURABLE_STATE_IS_CHANGED: 'MEASURABLE_STATE_IS_CHANGED',
-  MEASURABLE_STATE_IS_RESOLVED: 'MEASURABLE_STATE_IS_RESOLVED',
-  MEMBER_ADDED_TO_COMMUNITY: 'MEMBER_ADDED_TO_COMMUNITY',
-  NEW_INVITATION: 'NEW_INVITATION',
-  EMAIL_OUTER_TEMPLATE: 'EMAIL_OUTER_TEMPLATE',
-
-  MEMBER_JOINED_COMMUNITY_FEED_ITEM: 'MEMBER_JOINED_COMMUNITY_FEED_ITEM',
-  NEW_MEASUREMENT_PREDICTION_FEED_ITEM: 'NEW_MEASUREMENT_PREDICTION_FEED_ITEM',
-  NEW_MEASUREMENT_COMMENT_FEED_ITEM: 'NEW_MEASUREMENT_COMMENT_FEED_ITEM',
-  NEW_MEASUREMENT_RESOLUTION_FEED_ITEM: 'NEW_MEASUREMENT_RESOLUTION_FEED_ITEM',
-  NEW_MEASUREMENT_RESOLUTION_NOT_AVAILABLE_FEED_ITEM:
-    'NEW_MEASUREMENT_RESOLUTION_NOT_AVAILABLE_FEED_ITEM',
-  MEASURABLE_REACHED_RESOLUTION_DATE_FEED_ITEM:
-    'MEASURABLE_REACHED_RESOLUTION_DATE_FEED_ITEM',
-  NEW_MEASURABLE_FEED_ITEM: 'NEW_MEASURABLE_FEED_ITEM',
-  NEW_CHANNEL_FEED_ITEM: 'NEW_CHANNEL_FEED_ITEM',
-  NEW_NOTEBOOK_FEED_ITEM: 'NEW_NOTEBOOK_FEED_ITEM',
-};
-
-module.exports = {
-  TEMPLATE_NAME,
-};
