--- packages/server/src/enums/template-type.js
@@ -1,9 +1,0 @@
-const TEMPLATE_TYPE = {
-  EMAIL: 'EMAIL',
-  EMAIL_ENVELOPE: 'EMAIL_ENVELOPE',
-  FEED_ITEM: 'FEED_ITEM',
-};
-
-module.exports = {
-  TEMPLATE_TYPE,
-};
