--- packages/server/src/enums/token-type.js
@@ -1,8 +1,0 @@
-const TOKEN_TYPE = {
-  ACCESS_TOKEN: 'ACCESS_TOKEN',
-  AUTH_TOKEN: 'AUTH_TOKEN',
-};
-
-module.exports = {
-  TOKEN_TYPE,
-};
